Remove datasets.loader and add datasets.samplers (#2482)
* remove datasets.loader and add datasets.samplers

* fix typo in the filename
