--- mmdet/datasets/loader/__init__.py
@@ -1,4 +1,0 @@
-from .build_loader import build_dataloader
-from .sampler import DistributedGroupSampler, GroupSampler
-
-__all__ = ['GroupSampler', 'DistributedGroupSampler', 'build_dataloader']
