+++ mmdet/datasets/samplers/group_sampler.py
@@ -4,34 +4,7 @@ import math
 import numpy as np
 import torch
 from mmcv.runner import get_dist_info
-from torch.utils.data import DistributedSampler as _DistributedSampler
 from torch.utils.data import Sampler
-
-
-class DistributedSampler(_DistributedSampler):
-
-    def __init__(self, dataset, num_replicas=None, rank=None, shuffle=True):
-        super().__init__(dataset, num_replicas=num_replicas, rank=rank)
-        self.shuffle = shuffle
-
-    def __iter__(self):
-        # deterministically shuffle based on epoch
-        if self.shuffle:
-            g = torch.Generator()
-            g.manual_seed(self.epoch)
-            indices = torch.randperm(len(self.dataset), generator=g).tolist()
-        else:
-            indices = torch.arange(len(self.dataset)).tolist()
-
-        # add extra samples to make it evenly divisible
-        indices += indices[:(self.total_size - len(indices))]
-        assert len(indices) == self.total_size
-
-        # subsample
-        indices = indices[self.rank:self.total_size:self.num_replicas]
-        assert len(indices) == self.num_samples
-
-        return iter(indices)
 
 
 class GroupSampler(Sampler):
